Security update. HTML files are converted to PHP. The constant CORE is defined for access only to the allowed directories.

+++ cms/index.php
@@ -3,6 +3,8 @@
 // Connecting main classes
 require_once('_classes/index.php');
 
+if( !defined('CORE') ) { die('403'); }
+
 // Deleting the block
 if( isset($_GET['delete']) )
 	Content::deleteFieldAction($_GET['delete']);
